Use schedule to change state for motion sensor when switched off
The motion sensor rules were order-dependent, so it could happen that turning light off had no effect, as it was turned right on by another rule, since it seems that the bridge compares old value of the state. Instead, trigger a schedule, which sets motion sensor state blocked (-2) only after 1 second. At that time, it is safe and the other rule won't trigger.