Add /api/0/ prefix to command in the sensor off schedule
Normally, commands in schedules are for some unknown reasons prefixed via /api/<key>/, but the command for state config is too long this way.

Fortunately, any key can be specified to shorten the command and thus key "0" works as well. This fixes the problem with disabling the sensor before wake up timer.